t1/requirements.txt


aiohttp==3.8.5
asyncio==3.4.3
GitPython==3.1.31
langchain==0.0.184
langchain-pinecone==0.0.1
llama-index==0.7.7
octoai==0.1.3
pinecone-client==2.2.2
pydantic==1.10.7
python-dotenv==1.0.0
transformers==4.29.2